==> src/notebook-2-logit.Rmd <==
---
title: "Binary Logistic Regression"
author: "Ian Dinwoodie"
date: "`r Sys.Date()`"
output: html_document: default
---

```{r setup, include=FALSE}
library(tidyverse)
library(knitr)
library(broom)
library(plyr)
library(AER)
opts_chunk$set(echo=TRUE)
set.seed(1)
```

# Loading the Tidy Data

Load the tidy data from disk.

```{r}
df <- readRDS('../build/tidy.Rds')
#stopifnot(identical(dim(df)+0, c(1308, 261)))
str(df, list.len=5)
```

# Adjust the Data

TODO: Relocate this section to the tidy notebook.

```{r}
# Create dummy variables for the levels 1-4 of "behav_mod_strategy_used".
# We ignore levels 5 and 6 because they are irrelevant to the question.
df$has_systematic_desen <- ifelse(df$behav_mod_strategy_used==1, 1, 0)
df$has_systematic_desen[is.na(df$has_systematic_desen)] <- 0
df$has_systematic_desen <- as.logical(df$has_systematic_desen)
stopifnot(sum(df$has_systematic_desen) == 48)

df$has_counter_conditioning <- ifelse(df$behav_mod_strategy_used==2, 1, 0)
df$has_counter_conditioning[is.na(df$has_counter_conditioning)] <- 0
df$has_counter_conditioning <- as.logical(df$has_counter_conditioning)
stopifnot(sum(df$has_counter_conditioning) == 70)

df$has_sd_and_cc <- ifelse(df$behav_mod_strategy_used==3, 1, 0)
df$has_sd_and_cc[is.na(df$has_sd_and_cc)] <- 0
df$has_sd_and_cc <- as.logical(df$has_sd_and_cc)
stopifnot(sum(df$has_sd_and_cc) == 203)

df$has_leadership <- ifelse(df$behav_mod_strategy_used==4, 1, 0)
df$has_leadership[is.na(df$has_leadership)] <- 0
df$has_leadership <- as.logical(df$has_leadership)
stopifnot(sum(df$has_leadership) == 30)

df <- subset(df, select=-behav_mod_strategy_used)
#stopifnot(identical(dim(df)+0, c(1308, 269)))
```

```{r}
# Correct training method entries.
# We consider dogs trained with any form of punishment to be excluded from a
# reward-only training style. We consider compulsion a form of punishment.
df$has_punish_based_training <- ifelse(
    ((df$has_punish_based_training==1) | (df$has_combo_based_training==1) 
     | (df$has_compulsion_based_training==1)), 1, 0)
df$has_reward_based_training <- ifelse(df$has_punish_based_training==1, 0,
                                       df$has_reward_based_training)
df <- subset(df, select=-c(has_combo_based_training,
                           has_compulsion_based_training))
#stopifnot(identical(dim(df)+0, c(1308, 267)))

# Drop clicker/whistle from equipment since it is present in the behavior
# techniques.
df <- subset(df, select=-has_clicker)
#stopifnot(identical(dim(df)+0, c(1308, 266)))

# TODO: explain the lines below (68-71)
df$is_trainer_credentialed <- ifelse(df$is_trainer_credentialed == 1, 1, 0)
df[is.na(df$is_trainer_credentialed), 'is_trainer_credentialed'] <- 0
df$is_trainer_cpdt_credentialed <- ifelse(
    df$is_trainer_cpdt_credentialed == 1, 1, 0)
df[is.na(df$is_trainer_cpdt_credentialed), 'is_trainer_cpdt_credentialed'] <- 0

df$is_consultant_credentialed <- ifelse(
    df$is_consultant_credentialed == 1, 1, 0)
df[is.na(df$is_consultant_credentialed), 'is_consultant_credentialed'] <- 0
df$is_consultant_cbcc_ka <- ifelse(df$is_consultant_cbcc_ka == 1, 1, 0)
df$is_consultant_cdbc <- ifelse(df$is_consultant_cdbc == 1, 1, 0)
df$is_consultant_acaab <- ifelse(df$is_consultant_acaab == 1, 1, 0)
df$is_consultant_dacvb <- ifelse(df$is_consultant_dacvb == 1, 1, 0)
```

```{r}
# Condense punishing devices into a single column.
df$has_punishment_device <- ifelse(
    df$has_martingale == T, T, F)
df$has_punishment_device <- ifelse(
    df$has_prong_collar == T, T, df$has_punishment_device)
df$has_punishment_device <- ifelse(
    df$has_non_remote_electric_collar == T, T, df$has_punishment_device)
df$has_punishment_device <- ifelse(
    df$has_remote_electric_collar == T, T, df$has_punishment_device)
#df$has_punishment_device <- ifelse(
#    df$has_muzzle == T, T, df$has_punishment_device)
df <- subset(df,
             select=-c(has_martingale,
                       has_prong_collar,
                       has_non_remote_electric_collar,
                       has_remote_electric_collar,
                       has_muzzle))
str(df)
```

Retain only the columns necessary for analysis.

```{r}
# Retain only the columns to be used for analysis.
# - neuter status
# - age
# - breed <- we can't include this here because too many models
#            maybe we use the AKC breeding groups
# - aggression
# - age_of_acquisition
# - is_multi_dog_household

# Additional comments from Dr. Dodman:
# Age of onset for fear/anxiety is usually 0-2.
# You're less likely to get separation anxiety later on. However, there is a
# blip for geriatrics getting senior onset anxiety.
# Dogs usually develop serious thunderstorm phobia at 5-9 years

predictors <- c(
  '^is_male$',
  '^is_neutered$',
  '^has_seen_veterinarian$',
  '^has_seen_trainer$',
  '^has_seen_consultant$',
  #'^has_flat_collar$',
  #'^has_martingale$',
  #'^has_prong_collar$',
  #'^has_head_halter$',
  #'^has_harness_with_chest_lead$',
  #'^has_harness_with_shoulder_lead$',
  #'^has_non_remote_electric_collar$',
  #'^has_remote_electric_collar$',
  '^has_muzzle$',
  '^has_punishment_device$',
  '^has_other_training_equipment$',
  '^has_fluoxetine$',
  '^has_sertraline$',
  '^has_paroxetine$',
  '^has_citalopram$',
  '^has_escitalopram$',
  '^has_clomipramine$',
  '^has_trazodone$',
  '^has_buspirone$',
  '^has_valium$',
  '^has_clonidine$',
  '^has_acepromazine$',
  '^has_diphenhydramine$',
  '^has_other_medication$',
  '^has_acupuncture$',
  '^has_aromatherapy$',
  '^has_cannabidiol$',
  '^has_chiropracy$',
  '^has_dog_tv$',
  '^has_herbal_medicine$',
  '^has_homeopathy$',
  '^has_hormone_therapy$',
  '^has_hydrotherapy$',
  '^has_magnetic_field_therapy$',
  '^has_massage_therapy$',
  '^has_music_therapy$',
  '^has_nutraceuticals$',
  '^has_pheromone_therapy$',
  '^has_pressure_wraps$',
  '^has_reiki$',
  '^has_vision_blocking$',
  '^has_systematic_desen$',
  '^has_counter_conditioning$',
  '^has_sd_and_cc$',
  '^has_leadership$',
  '^has_mental_stimulation$',
  '^has_relaxation_protocol$',
  '^has_inc_physical_exercise$',
  '^has_habituation$',
  '^has_flooding$',
  '^has_response_blocking$',
  '^has_management$',
  '^has_red_of_stressors$',
  '^has_clicker_training$',
  '^has_play$',
  '^has_dog_sports$',
  '^has_ignoring_bad_behav$',
  '^has_training_alternatives$',
  '^has_short_freq_sessions$',
  '^has_group_classes$',
  '^has_private_sessions$',
  '^has_change_in_diet$',
  '^has_improved_comm$',
  '^has_other_behav_mod_technique$',
  '^has_exhibited_aggression$',
  '^has_reward_based_training$',
  '^is_trainer_credentialed$',
  '^is_trainer_cpdt_credentialed$',
  '^is_consultant_credentialed$',
  '^is_consultant_cbcc_ka$',
  '^is_consultant_cdbc$',
  '^is_consultant_acaab$',
  '^is_consultant_dacvb$'
)

# bins: {inanimate, animate, situational, generalized, ptsd}
outcomes <- c(
  '^fa_overall_rating$',
  '^inanimate_rating$',
  '^animate_rating$',
  '^situational_rating$',
  '^generalized_rating$',
  '^ptsd_rating$'
)
pattern <- paste(c(predictors, outcomes), collapse='|')
idx <- grep(pattern, names(df))
df <- df[, idx]
dim(df)
stopifnot(identical(dim(df)+0, c(1308, 75)))
```

All remaining columns except for the response to treatment for aggression scores
are factors. Update the column types to reflect this.

```{r}
for (col in names(df)) {
  if (col == 'fa_overall_rating') next
  if (col == 'inanimate_rating') next
  if (col == 'animate_rating') next
  if (col == 'situational_rating') next
  if (col == 'generalized_rating') next
  if (col == 'ptsd_rating') next
  df[, col] <- as.factor(df[, col])
}

str(df)
```

Get the number of dogs that used equipment:

```{r}
# TODO: Move to EDA notebook.
#df$used_equip <- ifelse(
#  (df$behavior_tech_used_1 == 1 | df$behavior_tech_used_2 == 1
#   | df$behavior_tech_used_3 == 1 | df$behavior_tech_used_4 == 1
#   | df$behavior_tech_used_5 == 1 | df$behavior_tech_used_6 == 1
#   | df$behavior_tech_used_7 == 1 | df$behavior_tech_used_8 == 1
#   | df$behavior_tech_used_9 == 1 | df$behavior_tech_used_10 == 1
#   | df$behavior_tech_used_11 == 1 | df$behavior_tech_used_12 == 1
#   | df$behavior_tech_used_13 == 1 | df$behavior_tech_used_14 == 1
#   | df$behavior_tech_used_15 == 1 | df$behavior_tech_used_16 == 1
#   | df$behavior_tech_used_17 == 1 | df$behavior_tech_used_18 == 1
#   | df$behavior_tech_used_19 == 1), TRUE, FALSE)
#has_mental_stimulation
#has_relaxation_protocol
#has_inc_physical_exercise
#has_habituation
#has_flooding
#has_response_blocking
#has_management
#has_red_of_stressors
#has_clicker_training
#has_play
#has_dog_sports
#has_ignoring_bad_behav
#has_training_alternatives
#has_short_freq_sessions
#has_group_classes
#has_private_sessions
#has_change_in_diet
#has_improved_comm
#has_other_behav_mod_technique
#
#summary(df$used_equip)
```

```{r}
# Drops NAs before regression
df <- df[!is.na(df$is_male),]
df <- df[!is.na(df$is_neutered),]
summary(df)
stopifnot(identical(dim(df)+0, c(1302, 75)))
```

General rule of thumb for logistic regression is that each variable should have
at least 10 responses per response option. We apply this rule of thumb below.

```{r}
# Drop the columns that do not meet the minimum response cutoff.
apply_binary_response_criteria <- function(df, min_cutoff=10)
{
  drops <- NULL
  outcomes <- NULL
  for (col in names(df)) {
    if (col == 'fa_overall_rating'
        || col == 'inanimate_rating'
        || col == 'animate_rating'
        || col == 'situational_rating'
        || col == 'generalized_rating'
        || col == 'ptsd_rating')
    {
        outcomes <- c(outcomes, col)
        next
    }
    
    counts <- count(df[, col])
    if (nrow(counts) < 2) {
      drops <- c(drops, col)
      next
    }
    
    for (row in 1:nrow(counts)) {
      if (counts[row, "freq"] < min_cutoff) {
        drops <- c(drops, col)
        break
      }
    }
  }
  
  return(df[, !(names(df) %in% drops)])
}

apply_min_xtab <- function(df, cutoff=10)
{
  drops <- NULL
  for (col in names(df)) {
    if (col == outcome) next
    xtab <- table(df[,col], df[,outcome])
    if (min(xtab) < cutoff) {
      drops <- c(drops, col)
      break
    }
  }
  
  print(drops)
  return(df[, !(names(df) %in% drops)])
}

df <- apply_binary_response_criteria(df)
stopifnot(identical(dim(df)+0, c(1302, 69)))
summary(df)
```

Now we initialize other functions that will be reused throughout the analysis.

```{r}
get_reduced_data_frame <- function(df, pred.patterns, outcome, excludes)
{
  pattern <- paste(c(pred.patterns, outcome), collapse='|')
  idx <- grep(pattern, names(df))
  df.out <- df[, idx]

  # Format the outcome as a factor with order.
  df.out[,outcome] <- factor(
      df.out[, outcome], levels=c("1","2","3","4","5","6","7"), ordered=TRUE)
  
  if (length(excludes) > 0) {
    # Exclude variables that cause issues for the model.
    df.out <- df.out[, !colnames(df.out) %in% excludes]
  }
  
  # Drop rows with null values for the outcome.
  df.out <- df.out[!is.na(df.out[, outcome]), ]
  
  return(df.out)
}
```

# Overall Fear/Anxiety

```{r fa_overall_df}
# Generate the desired subset data frame and collapse Likert outcome to be
# dichotomous.
outcome <- "fa_overall_rating"
excludes <- vector()
df_red <- apply_binary_response_criteria(
    get_reduced_data_frame(df, predictors, outcome, excludes))
df_red$fa_overall_rating <- as.factor(ifelse(df_red$fa_overall_rating <= 4, 0, 1))

summary(df_red)
```

```{r fa_overall_glm}
# Fit model to data.
f <- as.formula(paste0(outcome, "~", "."))
m <- glm(f, data=df_red, family="binomial")
summary(m)

print(vif(m))
```


```{r fa_overall_results, warning=FALSE}
df_results <- broom::tidy(m, conf.int=TRUE, exponentiate=TRUE)
#df_results$p.value <- p.adjust(df_results$p.value, method='BH')
df_results$sig <- ''
df_results[df_results$p.value <= .05, 'sig'] <- '*'
df_results[df_results$p.value <= .01, 'sig'] <- '**'
df_results[df_results$p.value <= .001, 'sig'] <- '***'
for (i in 1:nrow(df_results)) {
  if (is.na(df_results[i, 'conf.low']) | is.na(df_results[i, 'conf.high'])) next
  if ((df_results[i, 'conf.low'] < 1) & (df_results[i, 'conf.high'] > 1)) {
    df_results[i, 'sig'] <- ''
  }
}

knitr::kable(df_results)
```

# Inanimate Fear/Anxiety

```{r fa_inanimate_df}
# Generate the desired subset data frame and collapse Likert outcome to be
# dichotomous.
outcome <- 'inanimate_rating'
excludes <- vector()
df_red <- apply_binary_response_criteria(
    get_reduced_data_frame(df, predictors, outcome, excludes),
    min_cutoff=20)
df_red$inanimate_rating <- as.factor(
  ifelse(df_red$inanimate_rating <= 4, 0, 1))

summary(df_red)
```

```{r fa_inanimate_glm}
# Fit model to data.
f <- as.formula(paste0(outcome, "~", "."))
m <- glm(f, data=df_red, family="binomial")
summary(m)

print(vif(m))
```

```{r fa_inanimate_results, warning=FALSE}
df_results <- broom::tidy(m, conf.int=TRUE, exponentiate=TRUE)
#df_results$p.value <- p.adjust(df_results$p.value, method='BH')
df_results$sig <- ''
df_results[df_results$p.value <= .05, 'sig'] <- '*'
df_results[df_results$p.value <= .01, 'sig'] <- '**'
df_results[df_results$p.value <= .001, 'sig'] <- '***'
for (i in 1:nrow(df_results)) {
  if (is.na(df_results[i, 'conf.low']) | is.na(df_results[i, 'conf.high'])) next
  if ((df_results[i, 'conf.low'] < 1) & (df_results[i, 'conf.high'] > 1)) {
    df_results[i, 'sig'] <- ''
  }
}

knitr::kable(df_results)
```

# Animate Fear/Anxiety

```{r fa_animate_df}
# Generate the desired subset data frame and collapse Likert outcome to be
# dichotomous.
outcome <- "animate_rating"
excludes <- vector()

# Drop "inanimate_rating" to prevent pattern match with "animate_rating".
df_tmp <- subset(df, select=-inanimate_rating)
#stopifnot(identical(dim(df_tmp)+0, c(1296, 76)))

df_red <- apply_binary_response_criteria(
    get_reduced_data_frame(df_tmp, predictors, outcome, excludes),
    min_cutoff=10)
df_red$animate_rating <- as.factor(ifelse(df_red$animate_rating <= 4, 0, 1))

summary(df_red)
```

```{r fa_animate_glm}
# Fit model to data.
f <- as.formula(paste0(outcome, "~", "."))
m <- glm(f, data=df_red, family="binomial")
summary(m)

print(vif(m))
```

```{r fa_animate_results, warning=FALSE}
df_results <- broom::tidy(m, conf.int=TRUE, exponentiate=TRUE)
#df_results$p.value <- p.adjust(df_results$p.value, method='BH')
df_results$sig <- ''
df_results[df_results$p.value <= .05, 'sig'] <- '*'
df_results[df_results$p.value <= .01, 'sig'] <- '**'
df_results[df_results$p.value <= .001, 'sig'] <- '***'
for (i in 1:nrow(df_results)) {
  if (is.na(df_results[i, 'conf.low']) | is.na(df_results[i, 'conf.high'])) next
  if ((df_results[i, 'conf.low'] < 1) & (df_results[i, 'conf.high'] > 1)) {
    df_results[i, 'sig'] <- ''
  }
}

knitr::kable(df_results)
```

# Situational Fear/Anxiety

```{r fa_situational_df}
# Generate the desired subset data frame and collapse Likert outcome to be
# dichotomous.
outcome <- "situational_rating"
excludes <- vector()
df_red <- apply_binary_response_criteria(
    get_reduced_data_frame(df, predictors, outcome, excludes),
    min_cutoff=10)
df_red$situational_rating <-
    as.factor(ifelse(df_red$situational_rating <= 4, 0, 1))

summary(df_red)
```

```{r fa_situational_glm}
# Fit model to data.
f <- as.formula(paste0(outcome, "~", "."))
m <- glm(f, data=df_red, family="binomial")
summary(m)

print(vif(m))
```

```{r fa_situational_results, warning=FALSE}
df_results <- broom::tidy(m, conf.int=TRUE, exponentiate=TRUE)
#df_results$p.value <- p.adjust(df_results$p.value, method='BH')
df_results$sig <- ''
df_results[df_results$p.value <= .05, 'sig'] <- '*'
df_results[df_results$p.value <= .01, 'sig'] <- '**'
df_results[df_results$p.value <= .001, 'sig'] <- '***'
for (i in 1:nrow(df_results)) {
  if (is.na(df_results[i, 'conf.low']) | is.na(df_results[i, 'conf.high'])) next
  if ((df_results[i, 'conf.low'] < 1) & (df_results[i, 'conf.high'] > 1)) {
    df_results[i, 'sig'] <- ''
  }
}

knitr::kable(df_results)
```

# Generalized Fear/Anxiety

```{r fa_generalized_df}
# Generate the desired subset data frame and collapse Likert outcome to be
# dichotomous.
outcome <- "generalized_rating"
excludes <- vector()
df_red <- apply_binary_response_criteria(
    get_reduced_data_frame(df, predictors, outcome, excludes),
    min_cutoff=10)
df_red$generalized_rating <-
    as.factor(ifelse(df_red$generalized_rating <= 4, 0, 1))

summary(df_red)
```

```{r fa_generalized_glm}
# Fit model to data.
f <- as.formula(paste0(outcome, "~", "."))
m <- glm(f, data=df_red, family="binomial")
summary(m)

print(vif(m))
```

```{r fa_generalized_results, warning=FALSE}
df_results <- broom::tidy(m, conf.int=TRUE, exponentiate=TRUE)
#df_results$p.value <- p.adjust(df_results$p.value, method='BH')
df_results$sig <- ''
df_results[df_results$p.value <= .05, 'sig'] <- '*'
df_results[df_results$p.value <= .01, 'sig'] <- '**'
df_results[df_results$p.value <= .001, 'sig'] <- '***'
for (i in 1:nrow(df_results)) {
  if (is.na(df_results[i, 'conf.low']) | is.na(df_results[i, 'conf.high'])) next
  if ((df_results[i, 'conf.low'] < 1) & (df_results[i, 'conf.high'] > 1)) {
    df_results[i, 'sig'] <- ''
  }
}

knitr::kable(df_results)
```

# PTSD

```{r fa_ptsd_df}
# Generate the desired subset data frame and collapse Likert outcome to be
# dichotomous.
outcome <- "ptsd_rating"
excludes <- vector()
df_red <- apply_binary_response_criteria(
    get_reduced_data_frame(df, predictors, outcome, excludes),
    min_cutoff=20) # Increase cutoff due to smaller sample size.
df_red$ptsd_rating <- as.factor(ifelse(df_red$ptsd_rating <= 4, 0, 1))

# Collinearity adjustments.
df_red <- subset(df_red, select=-has_seen_consultant) # VIF 7.13

summary(df_red)
```

```{r fa_ptsd_glm}
# Fit model to data.
f <- as.formula(paste0(outcome, "~", "."))
m <- glm(f, data=df_red, family="binomial")
summary(m)

print(vif(m))
```

```{r fa_ptsd_results, warning=FALSE}
df_results <- broom::tidy(m, conf.int=TRUE, exponentiate=TRUE)
#df_results$p.value <- p.adjust(df_results$p.value, method='BH')
df_results$sig <- ''
df_results[df_results$p.value <= .05, 'sig'] <- '*'
df_results[df_results$p.value <= .01, 'sig'] <- '**'
df_results[df_results$p.value <= .001, 'sig'] <- '***'
for (i in 1:nrow(df_results)) {
  if (is.na(df_results[i, 'conf.low']) | is.na(df_results[i, 'conf.high'])) next
  if ((df_results[i, 'conf.low'] < 1) & (df_results[i, 'conf.high'] > 1)) {
    df_results[i, 'sig'] <- ''
  }
}

knitr::kable(df_results)
```

# Save Session Info

```{r}
sessionInfo()
```
